% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_functions.R
\name{IsDate}
\alias{IsDate}
\title{Check Date}
\usage{
IsDate(mydate, date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{mydate}{a sting that represent the date}

\item{date.format}{explicitly specifying the date format}
}
\value{
boolean value depending on the mydate value passed as argument
}
\description{
Function that checks the date format
}
\details{
This function checks if the date value entered is in the YYYY-MM-DD format
}
