% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_functions.R
\name{stock_details}
\alias{stock_details}
\title{Return Stock Data}
\usage{
stock_details(
  stock_name,
  start_date = Sys.Date() - lubridate::years(10),
  end_date = Sys.Date()
)
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date; By default 10 years from current date is considered}

\item{end_date}{a sting that represent the date; Current date is taken as default if no value is provided}
}
\value{
Stock Data Table
}
\description{
Function that return stock data based on the user inputs. The shiny app uses this function to get the data.
By default the function gives 10 years data for the stock if the start and end dates are not specified
}
\details{
The function initially cleans the entered argumets and check their validity. Then update the data file and finally
return the data table
}
\seealso{
\code{read.csv} \code{View}
}
\author{
Rijin Baby
}
