% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_functions.R
\name{stock_view}
\alias{stock_view}
\title{View Stock Plot}
\usage{
stock_view(
  stock_name,
  start_date = Sys.Date() - lubridate::years(10),
  end_date = Sys.Date()
)
}
\arguments{
\item{stock_name}{a sting that represent the symbol of the Company listed in NASDAQ}

\item{start_date}{a sting that represent the date; By default 10 years from current date is considered}

\item{end_date}{a sting that represent the date; Current date is taken as default if no value is provided}
}
\value{
Plot
}
\description{
Function that plot the stock data based on user input. By default the function gives 10 years data for the stock if the start and end dates are not specified
}
\details{
The function initially cleans the entered argumets and check their validity. Then update the data file and finally
dispalys the data for the analysis
}
\author{
Rijin Baby
}
