% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq_functions.R
\name{stock_view_app}
\alias{stock_view_app}
\title{Stock Analysis Application}
\usage{
stock_view_app()
}
\value{
Shiny Application
}
\description{
An user-friendly interactive application for historical stock price analysis.
}
\details{
Default screen shows AAPL data for 10 years. User can change the values on the side panel.
}
\author{
Rijin Baby
}
